streamlit/requirements.txt
streamlit
numpy
sounddevice
soundfile
groq
TTS
faster-whisper
python-dotenv